Added script to get beer list.
The script to get the beer list uses ruby to scrape the Frisco beer page based
on beer classes.  It then returns a JSON array of each beers' name and abv.  This uses the Mechanize Gem.  I don't know how to do this in Python.  If you can change it over, or just use this script.